feat: GetPlan 기능 추가

+++ src/main/java/com/example/planmate/entity/TimeTablePlaceBlock.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalTime;
 
 @Entity
@@ -25,7 +24,7 @@ public class TimeTablePlaceBlock {
     private String placeTheme;
 
     @Column(nullable = false, precision = 2, scale = 1)
-    private BigDecimal placeRating;
+    private float placeRating;
 
     @Column(nullable = false)
     private String placeAddress;
